requirements.txt
antlr4-python3-runtime==4.7.2
Pillow==10.1.0
pygraphviz==1.11
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
